+++ flow-typed/common/date.js
@@ -1,0 +1,4 @@
+declare type HumanizedDateTimeClasses = {
+  dateTime?: string,
+  humanizedText?: string,
+}