+++ src/date-input/index.js
@@ -1,4 +1,0 @@
-// @flow
-export { wrapPickerUtilProvider } from "./picker-util-provider-hoc";
-export { DateInput } from "./date-input";
-export { DateTimeInput } from "./date-time-input";