+++ src/index.js
@@ -12,7 +12,9 @@ export { CommonMultiSelect, CommonSelect } from "./common-select";
 export { CommonSuggest, SuggestionsOmniField } from "./common-suggest";
 export { CommonSwitch } from "./common-switch";
 export { CommonPage, CommonTabbedPage } from "./page";
-export { DateInput, DateTimeInput, wrapPickerUtilProvider } from "./date-input";
+export {
+  DateValue, DateTimeValue, HumanizedDateTime, DateInput, DateTimeInput, wrapPickerUtilProvider,
+} from "./date";
 export { ExternalLinkButton } from "./external-link-button";
 export { ExternalLink } from "./external-link";
 export { Navbar } from "./navbar";